Crate hyde_ipc

a wrapper around the hyprland library (used lovally for ipc)

Modules

config

This module provides helpers to easily config Hyprland

ctl

This module is for calling hyprctl commands, for getting data use data

data

This module provides functions for getting information on the compositor

dispatch

This module is for calling dispatchers and changing keywords

event_listener

This module provides the EventListener struct for listening and acting upon for events

keyword

This module provides the stuff needed to mutate, and read Hyprland config values

prelude

The prelude module, this is to import all traits

shared

This module provides shared things throughout the crate

Macros

async_closure

Creates a async closure

bind

Macro abstraction over Binder

command

Creates a CommandContent instance with the given flag and formatted data.

dispatch

Macro abstraction over Dispatch::call

Type Aliases

Result

This type provides the result type used everywhere in Hyprland-rs