hyde_ipc::ctl::set_prop

Enum PropType

pub enum PropType {

    AnimationStyle(String),
    Rounding(i64, bool),
    ForceNoBlur(bool, bool),
    ForceOpaque(bool, bool),
    ForceOpaqueOverriden(bool, bool),
    ForceAllowsInput(bool, bool),
    ForceNoAnims(bool, bool),
    ForceNoBorder(bool, bool),
    ForceNoShadow(bool, bool),
    WindowDanceCompat(bool, bool),
    NoMaxSize(bool, bool),
    DimAround(bool, bool),
    AlphaOverride(bool, bool),
    Alpha(f32, bool),
    AlphaInactiveOverride(bool, bool),
    AlphaInactive(f32, bool),
    ActiveBorderColor(Color, bool),
    InactiveBorderColor(Color, bool),

}

Type that represents a prop

Variants

AnimationStyle(String)

The animation style

Rounding(i64, bool)

The roundness

Tuple Fields

0: i64

1: bool

locked

ForceNoBlur(bool, bool)

Force no blur

Tuple Fields

0: bool

1: bool

locked

ForceOpaque(bool, bool)

Force opaque

Tuple Fields

0: bool

1: bool

locked

ForceOpaqueOverriden(bool, bool)

Force opaque overriden

Tuple Fields

0: bool

1: bool

locked

ForceAllowsInput(bool, bool)

Force allow input

Tuple Fields

0: bool

1: bool

locked

ForceNoAnims(bool, bool)

Force no animations

Tuple Fields

0: bool

1: bool

locked

ForceNoBorder(bool, bool)

Force no border

Tuple Fields

0: bool

1: bool

locked

ForceNoShadow(bool, bool)

Force no shadow

Tuple Fields

0: bool

1: bool

locked

WindowDanceCompat(bool, bool)

Allow for windoe dancing?

Tuple Fields

0: bool

1: bool

locked

NoMaxSize(bool, bool)

Allow for overstepping max size

Tuple Fields

0: bool

1: bool

locked

DimAround(bool, bool)

Dim around?

Tuple Fields

0: bool

1: bool

locked

AlphaOverride(bool, bool)

Makes the next setting be override instead of multiply

Tuple Fields

0: bool

1: bool

locked

Alpha(f32, bool)

The alpha

Tuple Fields

0: f32

1: bool

locked

AlphaInactiveOverride(bool, bool)

Makes the next setting be override instead of multiply

Tuple Fields

0: bool

1: bool

locked

AlphaInactive(f32, bool)

The alpha for inactive

Tuple Fields

0: f32

1: bool

locked

ActiveBorderColor(Color, bool)

The active border color

Tuple Fields

0: Color

1: bool

locked

InactiveBorderColor(Color, bool)

The inactive border color

Tuple Fields

0: Color

1: bool

locked