hyde_ipc::data

Struct Client

pub struct Client {

    pub address: Address,
    pub at: (i16, i16),
    pub size: (i16, i16),
    pub workspace: WorkspaceBasic,
    pub floating: bool,
    pub fullscreen: FullscreenMode,
    pub fullscreen_client: FullscreenMode,
    pub monitor: Option<i128>,
    pub initial_class: String,
    pub class: String,
    pub initial_title: String,
    pub title: String,
    pub pid: i32,
    pub xwayland: bool,
    pub pinned: bool,
    pub grouped: Vec<Box<Address>>,
    pub mapped: bool,
    pub swallowing: Option<Box<Address>>,
    pub focus_history_id: i8,

}

This struct holds information for a client/window

Fields
address: Address

The clientโs Address

at: (i16, i16)

The window location

size: (i16, i16)

The window size

workspace: WorkspaceBasic

The workspace its on

floating: bool

Is this window floating?

fullscreen: FullscreenMode

The internal fullscreen mode

fullscreen_client: FullscreenMode

The client fullscreen mode

monitor: Option<i128>

The monitor id the window is on, can be None in some cases

initial_class: String

The initial window class

class: String

The window class

initial_title: String

The initial window title

title: String

The window title

pid: i32

The process Id of the client

xwayland: bool

Is this window running under XWayland?

pinned: bool

Is this window pinned?

grouped: Vec<Box<Address>>

Group members

mapped: bool

Is this window print on screen

swallowing: Option<Box<Address>>

The swallowed window

focus_history_id: i8

When was this window last focused relatively to other windows? 0 for current, 1 previous, 2 previous before that, etc