hyde_ipc

Module dispatch

This module is for calling dispatchers and changing keywords

Dispatch module

This module is used for calling dispatchers and changing keywords

Usage

use hyprland::shared::HResult;
use hyprland::dispatch::{Dispatch, DispatchType};
fn main() -> HResult<()> {
   Dispatch::call(DispatchType::Exec("kitty"))?;

   Ok(())
}

Structs

Dispatch

The struct that provides all dispatching methods

Enums

Corner

This enum holds corners

CycleDirection

This enum holds a direction for cycling

Direction

This enum holds directions, typically used for moving

DispatchType

This enum holds every dispatcher

FocusMasterParam

Param for DispatchType::FocusMaster dispatcher

FullscreenType

This enum holds the fullscreen types

LockType

Enum used with DispatchType::LockGroups, to determine how to lock/unlock

MonitorIdentifier

This enum is used for identifying monitors

Position

This enum is used for resizing and moving windows precisely

SwapWithMasterParam

Param for DispatchType::SwapWithMaster dispatcher

WindowIdentifier

This enum is for identifying a window

WindowMove

This enum is the params to DispatchType::MoveWindow dispatcher

WindowSwitchDirection

This enum holds a direction for switch windows in a group

WorkspaceIdentifier

This enum is for identifying workspaces

WorkspaceIdentifierWithSpecial

This enum is for identifying workspaces that also includes the special workspace

WorkspaceOptions

This enum holds options that are applied to the current workspace