hyde_ipc

Module event_listener

This module provides the EventListener struct for listening and acting upon for events

Event Listener Module

for documentation go to:

crate::event_listener::EventStream for the event listener implementation based on the [futures_lite::Stream] api

crate::event_listener::EventListener for the normal Fn based event listener

crate::event_listener::AsyncEventListener for the Fn based event listener which uses closures that return std::future::Futures

Structs

AsyncEventListener

This struct is used for adding event handlers and executing them on events

ChangedSpecialEventData

This struct holds changed special event data

EventListener

This struct is used for adding event handlers and executing them on events

EventStream

Event listener, but [Stream] This is the new prefered way of listening for events as its more idiomatic, and allows for more efficient memory management

GroupToggledEventData

This struct holds the data for the Event::GroupToggled event

LayoutEvent

The data for the event executed when changing keyboard layouts

MonitorAddedEventData

This struct holds monitor event data

MonitorEventData

This struct holds monitor event data

NonSpecialWorkspaceEventData

This struct holds workspace event data when the workspace cannot be special

ScreencastEventData

Event data for screencast event

State

The mutable state available to Closures

UnknownEventData

This struct represents an unknown event to hyprland-rs this allows you to use events that haven’t been implemented in hyprland-rs. To use this use the UnknownEventData::parse_args method to properly get the args

WindowEventData

This struct holds window event data

WindowFloatEventData

This struct holds window float event data

WindowMoveEvent

The data for the event executed when moving a window to a new workspace

WindowOpenEvent

The data for the event executed when opening a new window

WindowPinEventData

This struct holds window pin event data

WindowTitleEventData

This struct holds the event data for the windowtitle changed event

WorkspaceEventData

This struct holds workspace event data

WorkspaceMovedEventData

This struct holds workspace moved event data

Enums

Event

This enum holds every event type