hyde_ipc

Module keyword

This module provides the stuff needed to mutate, and read Hyprland config values

Keyword module

This module is used for setting, and getting keywords

Usage

use hyprland::shared::HResult;
use hyprland::keyword::Keyword;
fn main() -> HResult<()> {
   Keyword::get("some_keyword")?;
   Keyword::set("another_keyword", "the value to set it to")?;

   Ok(())
}

Structs

Keyword

This struct holds a keyword

Enums

OptionValue

This enum holds the possible values of a keyword/option