hyde_ipc

Module shared

This module provides shared things throughout the crate

The Shared Module

This module provides shared private and public functions, structs, enum, and types

Macros

command

Creates a CommandContent instance with the given flag and formatted data.

Structs

Address

The address struct holds a address as a tuple with a single value and has methods to reveal the address in different data formats

CommandContent

This struct defines the content of a command, which consists of a flag and a data string.

Enums

CommandFlag

This enum defines the possible command flags that can be used.

HyprError

Error that unifies different error types used by Hyprland-rs

WorkspaceType

This enum holds workspace data

Traits

HyprData

This trait provides a standardized way to get data

HyprDataActive

Trait for helper functions to get the active of the implementor

HyprDataActiveOptional

Trait for helper functions to get the active of the implementor, but for optional ones

HyprDataVec

This trait provides a standardized way to get data in a from of a vector

Type Aliases

MonitorId

This type provides the id used to identify monitors

WorkspaceId

This type provides the id used to identify workspaces